+++ src/main/kotlin/org/coner/drs/ui/chooseevent/ChooseEventModel.kt
@@ -1,4 +1,4 @@
-package org.coner.drs.ui.choose_event
+package org.coner.drs.ui.chooseevent
 
 import io.reactivex.disposables.CompositeDisposable
 import javafx.beans.property.SimpleObjectProperty
